meta: Build both SPI and non-SPI firmware image
The full-size SPI image is only useful for boards that have SPI flash
installed (added or by default). For other boards that use MMC boot,
provide the unpadded firmware binary. To avoid confusion, add "spi" to
the name of the SPI flash image, and leave it in the board's build
directory.

Signed-off-by: Samuel Holland <samuel@sholland.org>