meta: Generate checksums for the firmware binaries
Since the point of the meta build system is to quickly generate firmware
images for several different boards for distribution, it makes sense to
generate sha256 and sha512 checksums in the output directory. Since the
checksum file depends on all built binaries, it allows simplifying the
`all` target as well.

Signed-off-by: Samuel Holland <samuel@sholland.org>