meta: Update U-Boot rule for binman conversion
Since sunxi moved from using mksunxi_fit_atf.sh to using binman,
u-boot.itb is no longer built, and u-boot-sunxi-with-spl.bin can no
longer be used as a make goal. Use the "all" goal to run binman, and
replace u-boot-sunxi-with-spl.bin with the new file having the same
contents: u-boot-sunxi-with-spl.fit.fit. The existence of this new
file is really a binman implementation detail, but it works for now.

Signed-off-by: Samuel Holland <samuel@sholland.org>