meta: Fix OUTDIR use in recipe that changes directory
U-Boot needs paths to the BL31 and SCP firmware binaries. The paths need
to be absolute, since the files are in OUTDIR, which could be anywhere
on the filesystem. Previously, this used CURDIR to generate an absolute
path. However, this breaks if BUILDDIR or OUTDIR is already absolute.
Instead, use a GNU Make function to generate the absolute path.

Signed-off-by: Samuel Holland <samuel@sholland.org>