meta: Do not keep a shared blank SPI image
Not all devices have the same SPI flash size, so sharing a blank image
between them can be problematic. The blank image does not get recreated
if FLASH_SIZE_KB is overridden on the command line.

The goal of using a blank image was to avoid unnecessary disk writes.
But copying the blank image to a temporary image writes the same amount
of data to disk as regenerating the image from nothing. So keep things
simple by removing the extra blank image file.

Signed-off-by: Samuel Holland <samuel@sholland.org>