meta: Include config files in the output directory
The build system does not enforce a specific configuration for Crust and
U-Boot. It applies a board's defconfig when BOARD changes (or during a
clean build), but the config files are not checked during later builds.
This is helpful when testing local configuration changes, but can cause
confusion if the configuration corresponding to the built binaries is
not recorded. Fix this by copying the config files to the board's output
directory.

Signed-off-by: Samuel Holland <samuel@sholland.org>