meta: Do not attempt to rebuild defconfigs
The Crust and U-Boot .config files depend on their respective defconfig
files as well as the source tree. If the source tree has not yet been
downloaded, the defconfig will not exist. Since the defconfig dependency
is processed first (due to appearing first in the list), it will fail to
resolve, and make errors out with:

  make: *** No rule to make target 'crust/.config', needed by
  'crust/build/scp/scp.bin'.  Stop.

Fix this by giving make an empty recipe for the defconfig file.

Signed-off-by: Samuel Holland <samuel@sholland.org>